Cambiado el Heap al Stack

+++ learning/cpp/myOCR/matrix.h
@@ -11,7 +11,7 @@
 namespace MyOCR {
 
     typedef unsigned char pixel;
-    typedef long int64;
+    typedef long long int int64;
 
     class Matrix {
         int width, height;
